outlook-ai/requirements.txt
annotated-types==0.7.0
certifi==2025.11.12
charset-normalizer==3.4.4
colorama==0.4.6
idna==3.11
markdown-it-py==4.0.0
mdurl==0.1.2
pip_system_certs==5.3
pydantic==2.12.4
pydantic_core==2.41.5
Pygments==2.19.2
python-dotenv==1.2.1
pywin32==311
requests==2.32.5
rich==14.2.0
tqdm==4.67.1
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.5.0
